pyproject.toml
[project]
name = "booking"
version = "1.0"
description = ""
authors = [
    {name = "YevgeniyKapustin",email = "kapustaevg@gmail.com"}
]
requires-python = ">=3.14"
dependencies = [
    "fastapi>=0.116.1",
    "uvicorn[standard] (>=0.40.0,<0.41.0)",
    "sqlalchemy>=2.0.43",
    "asyncpg>=0.30.0",
    "alembic>=1.16.5",
    "pydantic[email] (>=2.12.5,<3.0.0)",
    "python-jose>=3.5.0",
    "passlib[bcrypt] (>=1.7.4,<2.0.0)",
    "celery>=5.5.3",
    "redis>=6.4.0",
    "pytest>=8.4.1",
    "pydantic-settings (>=2.12.0,<3.0.0)",
    "gunicorn (>=25.0.2,<26.0.0)",
    "email-validator (>=2.3.0,<3.0.0)",
    "bcrypt==4.0.1",
    "prometheus-fastapi-instrumentator (>=7.0.0,<8.0.0)",
]

[tool.poetry]
package-mode = false

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "black (>=26.1.0,<27.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "isort (>=7.0.0,<8.0.0)",
    "mypy (>=1.19.1,<2.0.0)",
    "pytest-asyncio (>=0.25.0,<0.26.0)",
    "types-python-jose (>=3.5.0.20250531,<4.0.0.0)",
    "types-passlib (>=1.7.7.20250602,<2.0.0.0)",
    "aiosqlite (>=0.21.0,<0.22.0)"
]

[tool.mypy]
mypy_path = ["src"]
